function/requirements.txt
python-jose[cryptography]
requests
azure-storage-blob
azure-cosmos